+++ babel.config.js
@@ -1,7 +1,6 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
   plugins: [
-    ['transform-remove-console'],
     ['react-native-reanimated/plugin'],
     [
       require.resolve('babel-plugin-module-resolver'),
@@ -14,4 +13,11 @@ module.exports = {
       },
     ],
   ],
+  env: {
+    production: {
+      plugins: [
+        ['transform-remove-console'], // Remove console logs in production
+      ],
+    },
+  },
 };